Update fork-choice.md
Further mitigate bouncing attacks, by preventing more sophisticated ones which target the SAFE_SLOTS_TO_UPDATE_JUSTIFIED deadline to release the new justified checkpoint, so as to split validators into two groups, one which sees the new checkpoint in time and updates and one which does not. An appropriate split can prevent the current mitigation from working, and does not need a lot of stake to pull off
In addition to this change, it is also necessary to modify proposer behavior so that should_update_justified_checkpoint returns True at slot SAFE_SLOTS_TO_UPDATE_JUSTIFIED - 1 for the proposer of slot SAFE_SLOTS_TO_UPDATE_JUSTIFIED, i.e. the proposer of that slot gets one extra slot where they can update the justified checkpoint. This is a simple form of view-merge, allowing the proposer of slot SAFE_SLOTS_TO_UPDATE_JUSTIFIED to synchronize views on whether the checkpoint should be updated or not